Add Official Contest

+++ hoj-springboot/DataBackup/src/main/java/top/hcode/hoj/manager/admin/user/AdminUserManager.java
@@ -16,11 +16,15 @@ import org.springframework.transaction.annotation.Transactional;
 import org.springframework.util.StringUtils;
 import top.hcode.hoj.common.exception.StatusFailException;
 import top.hcode.hoj.dao.user.UserInfoEntityService;
+import top.hcode.hoj.dao.user.UserPreferencesEntityService;
 import top.hcode.hoj.dao.user.UserRecordEntityService;
+import top.hcode.hoj.dao.user.UserSignEntityService;
 import top.hcode.hoj.dao.user.UserRoleEntityService;
 import top.hcode.hoj.manager.msg.AdminNoticeManager;
 import top.hcode.hoj.pojo.dto.AdminEditUserDTO;
 import top.hcode.hoj.pojo.entity.user.UserInfo;
+import top.hcode.hoj.pojo.entity.user.UserPreferences;
+import top.hcode.hoj.pojo.entity.user.UserSign;
 import top.hcode.hoj.pojo.entity.user.UserRecord;
 import top.hcode.hoj.pojo.entity.user.UserRole;
 import top.hcode.hoj.pojo.vo.UserRolesVO;
@@ -51,6 +55,12 @@ public class AdminUserManager {
 
     @Autowired
     private UserRecordEntityService userRecordEntityService;
+
+    @Autowired
+    private UserSignEntityService userSignEntityService;
+
+    @Autowired
+    private UserPreferencesEntityService userPreferencesEntityService;
 
     @Autowired
     private RedisUtils redisUtils;
@@ -111,13 +121,18 @@ public class AdminUserManager {
         UpdateWrapper<UserInfo> userInfoUpdateWrapper = new UpdateWrapper<>();
         userInfoUpdateWrapper.eq("uuid", uid)
                 .set("username", username)
-                .set("realname", realname)
                 .set("email", email)
                 .set(setNewPwd, "password", SecureUtil.md5(password))
                 .set("title_name", titleName)
                 .set("title_color", titleColor)
                 .set("status", status);
         boolean updateUserInfo = userInfoEntityService.update(userInfoUpdateWrapper);
+
+        QueryWrapper<UserSign> userSignQueryWrapper = new QueryWrapper<>();
+        userSignQueryWrapper.eq("uid", uid);
+        UserSign userSign = userSignEntityService.getOne(userSignQueryWrapper, false);
+        userSign.setUsername(username).setRealname(realname);
+        boolean updateUserSign = userSignEntityService.updateById(userSign);
 
         QueryWrapper<UserRole> userRoleQueryWrapper = new QueryWrapper<>();
         userRoleQueryWrapper.eq("uid", uid);
@@ -133,7 +148,7 @@ public class AdminUserManager {
                 redisUtils.del(cacheKey);
             }
         }
-        if (updateUserInfo && setNewPwd) {
+        if (updateUserInfo && setNewPwd && updateUserSign) {
             // 需要重新登录
             userRoleEntityService.deleteCache(uid, true);
         } else if (changeUserRole) {
@@ -196,6 +211,9 @@ public class AdminUserManager {
     @Transactional(rollbackFor = Exception.class)
     public String addNewUser(List<String> user) throws StatusFailException {
         String uuid = IdUtil.simpleUUID();
+
+        UserSign userSign = new UserSign();
+        UserPreferences userPreferences = new UserPreferences();
         UserInfo userInfo = new UserInfo()
                 .setUuid(uuid)
                 .setUsername(user.get(0))
@@ -205,7 +223,7 @@ public class AdminUserManager {
         if (user.size() >= 4) {
             String realname = user.get(3);
             if (!StringUtils.isEmpty(realname)) {
-                userInfo.setRealname(user.get(3));
+                userSign.setRealname(user.get(3));
             }
         }
 
@@ -228,7 +246,7 @@ public class AdminUserManager {
         if (user.size() >= 7) {
             String school = user.get(6);
             if (!StringUtils.isEmpty(school)) {
-                userInfo.setSchool(school);
+                userSign.setSchool(school);
             }
         }
 
@@ -239,7 +257,10 @@ public class AdminUserManager {
         boolean result2 = userRoleEntityService.save(userRole);
         UserRecord userRecord = new UserRecord().setUid(uuid);
         boolean result3 = userRecordEntityService.save(userRecord);
-        if (!result1 || !result2 || !result3) {
+        boolean result4 = userSignEntityService.save(userSign);
+        boolean result5 = userPreferencesEntityService.save(userPreferences);
+
+        if (!result1 || !result2 || !result3 || !result4 || !result5) {
             throw new StatusFailException("生成用户失败");
         }
         return uuid;
@@ -253,15 +274,18 @@ public class AdminUserManager {
         int numberFrom = (int) params.getOrDefault("number_from", 1);
         int numberTo = (int) params.getOrDefault("number_to", 10);
         int passwordLength = (int) params.getOrDefault("password_length", 6);
+        String passwd = (String) params.getOrDefault("password_custom", "");
 
         List<UserInfo> userInfoList = new LinkedList<>();
         List<UserRole> userRoleList = new LinkedList<>();
         List<UserRecord> userRecordList = new LinkedList<>();
+        List<UserSign> userSignList = new LinkedList<>();
+        List<UserPreferences> userPreferencesList = new LinkedList<>();
 
         HashMap<String, Object> userInfo = new HashMap<>(); // 存储账号密码放入redis中，等待导出excel
         for (int num = numberFrom; num <= numberTo; num++) {
             String uuid = IdUtil.simpleUUID();
-            String password = RandomUtil.randomString(passwordLength);
+            String password = StringUtils.isEmpty(passwd) ? RandomUtil.randomString(passwordLength) : passwd;
             String username = prefix + num + suffix;
             userInfoList.add(new UserInfo()
                     .setUuid(uuid)
@@ -272,11 +296,16 @@ public class AdminUserManager {
                     .setRoleId(1002L)
                     .setUid(uuid));
             userRecordList.add(new UserRecord().setUid(uuid));
+            userSignList.add(new UserSign().setUid(uuid).setUsername(username));
+            userPreferencesList.add(new UserPreferences().setUid(uuid));
         }
         boolean result1 = userInfoEntityService.saveBatch(userInfoList);
         boolean result2 = userRoleEntityService.saveBatch(userRoleList);
         boolean result3 = userRecordEntityService.saveBatch(userRecordList);
-        if (result1 && result2 && result3) {
+        boolean result4 = userSignEntityService.saveBatch(userSignList);
+        boolean result5 = userPreferencesEntityService.saveBatch(userPreferencesList);
+
+        if (result1 && result2 && result3 && result4 && result5) {
             String key = IdUtil.simpleUUID();
             redisUtils.hmset(key, userInfo, 1800); // 存储半小时
             // 异步同步系统通知
